src/main/java/chess/domain/chessPiece/pieceType/Queen.java
package chess.domain.chessPiece.pieceType;

import chess.domain.chessPiece.ChessPiece;
import chess.domain.chessPiece.pieceState.PieceState;

public class Queen extends ChessPiece {

	public static final String NAME = "Q";
	private static final int SCORE = 9;

	public Queen(PieceColor pieceColor, PieceState pieceState) {
		super(pieceColor, pieceState);
	}

	@Override
	public String getName() {
		return pieceColor.convertFrom(NAME);
	}

	@Override
	public double getScore() {
		return SCORE;
	}

}
